pyproject.toml
[project]
name = "ecommerce-triage-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.0.22",
    "livekit-plugins-cartesia>=1.0.22",
    "livekit-plugins-deepgram>=1.0.22",
    "livekit-plugins-noise-cancellation>=0.2.4",
    "livekit-plugins-openai>=1.0.22",
    "livekit-plugins-silero>=1.0.22",
    "python-dotenv~=1.0",
    "pyyaml>=6.0.2",
]
